-#355: Refactor SharedFromMe
